fix: test and type issues

+++ database/migrations/2025_11_27_001342_create_taxes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->string('tax_type');
-            $table->decimal('rate', 8, 2);
+            $table->unsignedBigInteger('rate');
             $table->boolean('is_active')->index();
 
             $table->foreignId('created_by')->references('id')->on('users');
